requirements.txt
langchain
langchain-openai
langchain-community
openai
chromadb
tiktoken
python-dotenv
